Mods to support 64 bit

+++ src/yacam.cpp
@@ -400,7 +400,7 @@ RC YACAM::vprintf(		// "vprintf" to file
 {
 	char buf[514];
 	vsprintf( buf, fmt, ap);				// format arguments (if any) into given format string
-	return putBytes( buf, strlen(buf), -1L, erOp);	// write (above), return result
+	return putBytes( buf, static_cast<int>(strlen(buf)), -1L, erOp);	// write (above), return result
 }
 //===========================================================================
 //  YACAM character sequential read functions: peekC, getC, toke, get.
@@ -812,7 +812,7 @@ int YACAM::scanLine(			// scan file for line match
 	int bBeg = (erOp & EROP1) != 0;
 	char tS[ 1000];
 	strTrim( tS, s);		// trimmed s
-	int tSLen = strlen( tS);
+	int tSLen = static_cast<int>(strlen(tS));
 	int ret = 0;
 	int iL;
 	for (iL=1; ret==0; iL++)
